+++ mail/mail-uniffi/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [ET-1864] `account_details` function in `MailUserSession` and `details` function in `StoredAccount` (it contains account name, email and avatar information that needs to be displayed to the user).
   - [ET-1794] `DecryptedMessage::get_attachments` which merges the API attachments and PGP attachments into one for easier client consumption.
   - [ET-1685] `scroll_search` & `SearchScroller` to make server searches
+  - [ET-1385] Resolve contact group total for message recipients.
 
 ### Changed